Check if table exists before dropping

+++ db/migrate/20140624143329_drop_os_tags.rb
@@ -1,6 +1,6 @@
 class DropOsTags < ActiveRecord::Migration
   def up
-    drop_table :os_tags
+    drop_table :os_tags if ActiveRecord::Base.connection.table_exists? :os_tags
   end
 
   def down
